apply review comment
