+++ app/a-propos/content.mdx
@@ -1,6 +1,5 @@
 import { Socials } from '../../modules/about/Socials';
 import { Orgas } from '../../modules/about/Orgas';
-import { Numbers } from '../../modules/about/Numbers';
 
 <main>
 
@@ -25,10 +24,6 @@ import { Numbers } from '../../modules/about/Numbers';
 
     Nous organisons également de manière régulière des ApérosJS, qui n'ont pour unique but que de rassembler
     la communauté dans un lieu lyonnais 🍻
-
-    ## Quelques chiffres
-
-    <Numbers />
 
     ## Les orgas
