+++ app/page.tsx
@@ -3,6 +3,7 @@ import { HomeHero } from '../modules/home/HomeHero';
 import { SeePastEvents } from '../modules/home/Home.components';
 import { Sponsors } from '../modules/sponsors/Sponsors';
 import { NextEvent } from '../modules/home/NextEvent';
+import { Numbers } from '../modules/home/Numbers';
 
 export const revalidate = 3600;
 export default function Home() {
@@ -11,6 +12,7 @@ export default function Home() {
       <HomeHero />
       <NextEvent />
       <SeePastEvents />
+      <Numbers />
       <Sponsors />
     </main>
   );
